schema: use the index type from the schema and not from the base table
The describe schema method of a view needs to know if it's an index and
if it's a global index.

This patch change the way the method is getting the information, instead
of taking it from the base table (that can be unavailable) it takes it
from the schema.

Fixes #5614

Signed-off-by: Amnon Heiman <amnon@scylladb.com>